* src/space.c (find_readonly_space): Disallow section insertion in 	between adjacent SHT_NOTE sections.  When creating new PT_LOAD, 	handle multiple consecutive SHT_NOTE sections with just one PT_NOTE 	segment for them. 	* testsuite/Makefile.am (TESTS): Add shuffle9.sh. 	* testsuite/Makefile.in: Regenerated. 	* testsuite/shuffle9.sh: New test. 	* testsuite/shuffle9.c: New.
	* testsuite/shuffle2.c: Add .previous at the end of __asm.
	* testsuite/reloc5.c (main): Likewise.

	* testsuite/tls3.sh: If libs aren't built with -fpic,
	disable the test under SELinux in enforcing mode.


git-svn-id: svn://sourceware.org/svn/prelink/trunk@156 94c539fb-cf18-0410-b60f-edeeb537fa16

+++ testsuite/tls3.sh
@@ -9,6 +9,8 @@ SHFLAGS=
 case "`uname -m`" in
   ia64|ppc*|x86_64|alpha*|s390*|mips*) SHFLAGS=-fpic;; # Does not support non-pic shared libs
 esac
+# Disable this test under SELinux if textrel
+test -z "$SHFLAGS" -a -x /usr/sbin/getenforce -a "`/usr/sbin/getenforce`" = Enforcing && exit 77
 rm -f tls3 tls3lib*.so tls3.log
 rm -f prelink.cache
 $CC -shared -O2 -fpic -o tls3lib1.so $srcdir/tls1lib1.c
